update stitch, fix error when mappings partially present

+++ src/main/java/net/fabricmc/loom/providers/PomfProvider.java
@@ -71,10 +71,16 @@ public class PomfProvider extends DependencyProvider {
 		}
 
 		if (!MAPPINGS_TINY_BASE.exists() || !MAPPINGS_TINY.exists()) {
-			project.getLogger().lifecycle(":extracting " + mappingsJar.getName());
-			try (FileSystem fileSystem = FileSystems.newFileSystem(mappingsJar.toPath(), null)) {
-				Path fileToExtract = fileSystem.getPath("mappings/mappings.tiny");
-				Files.copy(fileToExtract, MAPPINGS_TINY_BASE.toPath());
+			if (!MAPPINGS_TINY_BASE.exists()) {
+				project.getLogger().lifecycle(":extracting " + mappingsJar.getName());
+				try (FileSystem fileSystem = FileSystems.newFileSystem(mappingsJar.toPath(), null)) {
+					Path fileToExtract = fileSystem.getPath("mappings/mappings.tiny");
+					Files.copy(fileToExtract, MAPPINGS_TINY_BASE.toPath());
+				}
+			}
+
+			if (MAPPINGS_TINY.exists()) {
+				MAPPINGS_TINY.delete();
 			}
 
 			project.getLogger().lifecycle(":populating field names");
